+++ package.json
@@ -49,12 +49,10 @@
   "devDependencies": {
     "@commitlint/cli": "18.6.0",
     "@commitlint/config-conventional": "18.6.0",
-    "@faker-js/faker": "8.4.0",
     "@trivago/prettier-plugin-sort-imports": "4.3.0",
     "@types/eslint": "8.56.2",
     "@types/js-cookie": "3.0.6",
     "@types/jsdom": "21.1.6",
-    "@types/lodash.merge": "^4",
     "@types/node": "20.11.14",
     "@types/uuid": "9.0.8",
     "@types/validator": "13.11.8",
